Defer counting on paginated datasets
This is a (failed) experiment to determine if the pagination extension can defer the `SELECT COUNT` it does until it’s needed by the paginated dataset. This change makes the specs fail in a way which demonstrates why it can’t: the dataset is frozen, so it needs the two counts when it’s created.